+++ R/sm.R
@@ -19,8 +19,7 @@
 #' * `dfzb`: the features scaled by coefficients of the Ridge Regression model
 #' @keywords public
 #' @examples \donttest{
-#' RPsub <- RP[1:50, ]
-#' x <- selective_measuring(RPsub, k = 5, verbose = 0)
+#' x <- selective_measuring(RP[1:50, ], k = 5, verbose = 0)
 #' # For the sake of a short runtime, only the first 50 rows of the RP dataset
 #' # were used in this example. In practice, you should always use the entire
 #' # dataset to find the optimal subset for re-measurement.